memory: Support enabling memory profiling at runtime
Changes the heap profiler code such that there is now two levels of
enablement:

 - First the SEASTAR_HEAPPROF define continues to exist for total
   removal of the profiling code.
 - Second additional runtime options: This allows enabling profiling at
   runtime.

The only functional change is to switch to `simple_backtrace` from
`tasktrace`. The added `tasktrace` functionality is not used in the gdb
scripts so this reduces overhead associated with that.

There is no other functional changes to the heap profiling. It's mostly
just moving a bunch of code around and adding some if statements.

It's not particularly useful for now as one has to be careful to not
deallocate non-sampled allocations while sampling is on.

This is a step in preparation for actual sampled heap profiling which
will allow us to differentiate between sampled and non-sampled
allocations and also to not profile every single allocation.

Also add a very basic test we can expand on.