memory: Add allocator perf tests
Adds some very basic micro benchmarks

They are mostly focused on single allocations for now with the idea that
they will help us to confirm that non-sampled allocations are not
affected too much by the additional overhead if sampling is turned on.

Also they are not "system benchmarks", i.e.: they don't try to create
random load to somehow simulate a real workload.