iotune: Check the disk for duplex-ness
Modern SSDs are full duplex (i.e.: they can support full write and read
performance at the same time) and the io-queue already has support for
this.

However, iotune had no way of actually producing io-properties that set
this.

This patch adds an additional test to iotune which re-runs the write and
read IOPS tests but at the same time. If we achieve at least 90% of our
previously separately measured IOPS we set the duplex flag.