iotune: Ignore measurements during warmup period
Make iotune ignore the first three seconds of measurements. Disks are
often fairly bursty in the initial part of the test. This skews the
measured measurements and makes us produce io-properties that are
potential way too large.

It also heavily affects the shown deviation which can look a lot worse
than it is.